requirements/base.txt
Django==5.1.*
djangorestframework==3.15.*
django-filter==24.*
dj-database-url==2.*
celery[redis]==5.4.*
redis==5.*
psycopg[binary]==3.*
boto3==1.35.*
pandas==2.2.*
pyarrow==17.*
pulp==2.8.*
scikit-learn==1.5.*
numpy==1.26.*
plotly==5.24.*
